Add variant prop to deal with wrappers
