+++ docs/pages/api/datepicker/index.mdx
@@ -46,4 +46,4 @@ It is allowed to hardly customize displaying of dates. Thus you can add badges o
 
 <Example source={CustomDay} />
 
-<PropTypesTable src="DatePicker/DatePickerModal.tsx" />
+<PropTypesTable src="DatePicker/DatePicker.tsx" />
